pyproject.toml
[project]
name = "dtcwt-video"
version = "0.1.0"
description = "3D DT-CWT 기반 비디오 전처리 연구: Noisy & Low-bitrate 환경에서의 조건부 RD 성능 개선 분석"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "dtcwt>=0.14.0",
    "ffmpeg-python>=0.2.0",
    "matplotlib>=3.10.8",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pyopencl>=2026.1.2",
    "pywavelets>=1.7.0",
    "torch>=2.0.0",
    "y4m>=1.1.1",
    "yuvio>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
dtcwt-noise = "scripts.run_noise_experiment:main"
dtcwt-rd    = "scripts.run_rd_curve:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dtcwt_video", "scripts"]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "requires_ffmpeg: 테스트에 FFmpeg(libvmaf 포함) 또는 libx264가 필요함",
    "requires_gpu: 테스트에 CUDA GPU가 필요함",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]
